backend/requirements.txt
fastapi
uvicorn[standard]
websockets
httpx
python-dotenv
mistralai
google-genai
Pillow
